Move remaining R2 files into r2 package
